Automation keeping presales fork in sync with dc-demostore-core
This automation should only run on the downstream fork of this codebase. Keeping the downstream fork (dc-demostore-presales) in-sync with changes to the parent (dc-demostore-core).

Making this change in the parent repo, prevent's a merge conflict in the downstream fork by adding this file there.

This automation contains logic to prevent this pipeline step from running on the parent repo, or running on any other fork of this repo except dc-demostore-presales. This shouldn't pose a risk to other forks / consumers of this codebase.

This has been done manually in the past, and can be automated like this.